Tests/LibWeb/Text/expected/anchor-scroll-shift-with-anchor-after-positioned-box.txt

top gap before scroll: 0
left gap before scroll: 0
anchor bottom moved by: -40
top gap after scroll: 0
left gap after scroll: 0
